% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{TSNEintgraph}
\alias{TSNEintgraph}
\title{TSNEintgraph needs to be tested}
\usage{
TSNEintgraph(intGraph, seuratObj, metaCol, metaColPlot = NULL, numberCut,
  numberSplit)
}
\arguments{
\item{intGraph}{an output from intgraph}

\item{seuratObj}{a Seurat object}

\item{metaCol}{character, name of metadata column corresponding to nodes of intGraph}

\item{metaColPlot}{name of metadata column corresponding to nodes of intGraph if different from metaCol, default NULL}

\item{numberCut}{numeric value for minimum number of interactions to plot (default 20)}

\item{numberSplit}{numeric value for spliting numbers of interactions (i.e high vs low, default 60)}
}
\value{
a plot
}
\description{
building on intgraph, adds network to a TSNE plor
}
